+++ tests/Unit/Doctrine/Phpcr/IdPrefixListenerTest.php
@@ -41,7 +41,7 @@ class IdPrefixListenerTest extends TestCase
      */
     protected $routeMock;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->candidatesMock = $this->createMock(PrefixCandidates::class);
         $this->candidatesMock